Support v2 of Pager Duty REST API
This commit updates the Connection class to support only v2+ of
the Pager Duty REST API. Updates have been made following the migration
document https://v2.developer.pagerduty.com/docs/migrating-to-api-v2.
Notable v2 updates include:

- Required 'Accept' request header which includes version number
- Base URL has changed, so account_domain is no longer needed
- Rate limiting error handling
- requester_id (user id) is no longer sent as a query parameter to
identify the requesting user. Instead, set the From header to the
request user's email address in the headers hash as needed.

In addition to these v2 updates, method signatures have been changed, which
was the main motivation for no longer supporting v1. Prior to
this update, the query parameters were sent through the request body,
which worked for v1 but not for v2.

One of the changes made to v2 was to
restructure the way lists are sent as query parameters. Instead of
having test.com/some_uri?status=status1,status2,status3, we now need to
support
test.com/some_url?statuses[]=status1&statuses[]=status2&statuses[]=status3. Faraday supports encoding query parameters this way, but when sent through the body, arrays are encoded as JSON and so they do not align with what the API requires. Additionally, now that requester_id is added as a header value, it makes sense to separate the request body, headers, and query parameters.

Additionally, we no longer default to any version of the API. The API
version should be sent when initializing PagerDuty::Connection.

[151816727386211](https://app.asana.com/0/151816727386211/151816727386211)