Fix raise on non-200 status code
The execution order of Faraday::Middleware passed to conn.use
in the Faraday.new block is the reverse order of the calls to
conn.use (i.e. if conn.use Klass comes last it will run first).
Because the time string parsing middleware raises on the body
when the status code is 401, and because ParseTimeStrings
executes before RaiseApiErrorOnNon200, the raise for a non-200
status code never happens.

Moving the error status code handling middleware to the end of
the block ensures that they execute first.

Fixes technicalpickles/pager_duty-connection#17.