Refactor `PagerDuty::Connection#get` so it doesn't mutate the input
This refactors `PagerDuty::Connection#get` so that it doesn't
mutate the `request` passed to it.

At the moment, it mutates the `:query_params` key inside the
`request`, potentially dropping the `:page` and `:limit`
keys (due to use of `Hash#delete`) and adding a calculated
`:offset`.

Mutating the incoming value isn't a problem - but it's certainly
a bit surprising, and an issue that a user is likely to run into
when paginating through records with a Ruby `loop`.